More Action Pack `abstract_unit` cleanup
- Remove dead classes / dead code
- Move class definitions to where they are used, don't define in a
  shared space
